refactor: extract shared four-eyes exception mappers into libs-runtime
SelfApprovalNotAllowedMapper/InvalidApprovalStateMapper (403/409 for the
ADR-0155 four-eyes exceptions) were duplicated byte-for-byte across 10
services (account, balance, clearing, domestic-payment, ledger, lending,
sepa-instant, sepa-payment, swift, transaction), plus a divergent
{"code","message"}-shaped copy in notification-service. Move them into
openbank-libs-runtime's already-auto-registered CommonExceptionMappers.kt
and delete every local copy.

notification-service is included: once these mappers are auto-registered
fleet-wide via Jandex, leaving its local override in place would register
two ExceptionMapper<SelfApprovalNotAllowedException> providers for the
same service — the exact non-deterministic JAX-RS provider-selection bug
(issue #526) this repo already guards against for JDK types, just for a
domain type check-exception-mapper-collision.sh didn't cover yet. That
script's guarded type list is extended accordingly.

Closes #1394
