+++ openbank-ledger-service/src/main/kotlin/com/openbank/ledger/infrastructure/rest/ExceptionMappers.kt
@@ -12,10 +12,6 @@ import com.openbank.ledger.application.usecase.YearCloseConflictException
 import com.openbank.ledger.application.usecase.YearCloseNotFoundException
 import com.openbank.ledger.domain.model.LedgerConflictException
 import com.openbank.ledger.domain.model.LedgerValidationException
-import com.openbank.libs.api.error.ApiError
-import com.openbank.libs.approval.InvalidApprovalStateException
-import com.openbank.libs.approval.SelfApprovalNotAllowedException
-import com.openbank.libs.domain.identifiers.Ids
 import jakarta.ws.rs.core.MediaType
 import jakarta.ws.rs.core.Response
 import jakarta.ws.rs.core.Response.Status.CONFLICT
@@ -109,42 +105,9 @@ class ClosedFiscalPeriodExceptionMapper : ExceptionMapper<ClosedFiscalPeriodExce
         .build()
 }
 
-// ADR-0155: a checker can never decide their own PendingApproval — ApprovalStore.decide
-// enforces this itself (defense-in-depth), surfaced here as a plain 403.
-@Provider
-class SelfApprovalNotAllowedMapper : ExceptionMapper<SelfApprovalNotAllowedException> {
-    override fun toResponse(exception: SelfApprovalNotAllowedException): Response =
-        Response.status(Response.Status.FORBIDDEN)
-            .entity(
-                ApiError(
-                    // ADR-0106: a per-response correlation id, not a durable/indexed identifier — Ids.randomId().
-                    traceId = Ids.randomId().toString(),
-                    status = 403,
-                    code = "FORBIDDEN",
-                    message = exception.message ?: "Forbidden",
-                ),
-            )
-            .build()
-}
-
-// decide()/markExecuted() reject re-deciding or re-consuming an approval that isn't in the
-// expected status (guards against an EXECUTED approval being flipped back to APPROVED and
-// replayed). Surfaced as a 409, matching the existing conflict-mapper convention above for
-// state-machine violations.
-@Provider
-class InvalidApprovalStateMapper : ExceptionMapper<InvalidApprovalStateException> {
-    override fun toResponse(exception: InvalidApprovalStateException): Response =
-        Response.status(Response.Status.CONFLICT)
-            .entity(
-                ApiError(
-                    traceId = Ids.randomId().toString(),
-                    status = 409,
-                    code = "CONFLICT",
-                    message = exception.message ?: "Conflict",
-                ),
-            )
-            .build()
-}
+// SelfApprovalNotAllowedMapper / InvalidApprovalStateMapper (403/409) moved to
+// openbank-libs-runtime's CommonExceptionMappers (issue #1394) — a service-local copy of the
+// same exact type would collide non-deterministically with the shared one (issue #526).
 
 // ExceptionMapper<Exception> (GlobalExceptionMapper) is intentionally NOT declared here.
 // openbank-libs auto-registers GenericExceptionMapper (Exception → 500, correlation-aware) and
